refactor: move hal pki → hal vault pki with --k8s flag
- Delete cmd/pki/ package (6 files)
- Create cmd/vault/pki.go in the vault package
  - hal vault pki [status|enable|disable|update]
  - --k8s flag on enable/update triggers cert-manager + web demo
  - --acme reserved for future ACME challenge support
  - disable auto-detects and cleans cert-manager regardless of --k8s flag
  - dedicated kubernetes-pki/ auth mount, never shared with kubernetes/
  - conditional KinD deletion (preserved if hal vault k8s is still active)
- Remove hal/cmd/pki import and AddCommand from cmd/root.go
- Rewrite docs/commands/vault-pki.md for new command surface
- Update README.md: add hal vault pki section
